Metrics - more casts from numeric to int8 (minor)

+++ pgwatch2/metrics/replication_slots/10/metric.sql
@@ -4,6 +4,6 @@ select
   coalesce(plugin, 'physical')::text as tag_plugin,
   active,
   case when active then 0 else 1 end as non_active_int,
-  pg_wal_lsn_diff(pg_current_wal_lsn(), restart_lsn) as restart_lsn_lag_b
+  pg_wal_lsn_diff(pg_current_wal_lsn(), restart_lsn)::int8 as restart_lsn_lag_b
 from
   pg_replication_slots;
